+++ frontend/src/login/forgotPassword/ForgotPasswordModal.tsx
@@ -16,8 +16,8 @@ If not, see <https://www.gnu.org/licenses/>.
 */
 
 import React, { Component } from 'react'
-import Modal, { IModalProps } from './Modal'
-import Email from '../form/EmailAddress';
+import Modal, { IModalProps } from '../../shared/components/Modal'
+import Email from '../../shared/form/EmailAddress';
 import Button from '@material-ui/core/Button';
 import './ForgotPasswordModal.css';
